convert item dependency tree to list
