+++ frontend/src/features/archived-items/item-dependency-list/index.ts
@@ -1,0 +1,1 @@
+import "./item-dependency-list";
