+++ frontend/src/features/archived-items/item-dependency-list/item-dependency-list.stylesheet.css
