--- frontend/src/features/archived-items/item-dependency-tree/index.ts
@@ -1,1 +1,0 @@
-import "./item-dependency-tree";
