+++ frontend/src/stories/features/collections/DedupeWorkflows.ts
@@ -3,7 +3,7 @@ import { html } from "lit";
 import type { DedupeWorkflows } from "@/features/collections/dedupe-workflows/dedupe-workflows";
 
 import "@/features/collections/dedupe-workflows/dedupe-workflows";
-import "@/features/archived-items/item-dependency-tree";
+import "@/features/archived-items/item-dependency-list";
 import "@/features/archived-items/crawl-status";
 
 export type RenderProps = DedupeWorkflows;
